requirements.txt
spacy==3.7.6
beautifulsoup4==4.12.3
markdownify==0.13.1
requests
openai==1.46.0
sentence-transformers==3.1.0
pprintpp
numpy
lxml
pyyaml
rank-bm25
